Change GHA PR workflows to checkout HEAD of the PR
By default it uses GITHUB_REF as the ref, which is the merge commit
hash for PRs.
The reason for this change is that we had a PR that when merged would
fail linting check while being confusing, as the linting passed on the
branch which was behind the target branch.